Install mocha & chai test suite
- Adds tests for events, along with scaffolding for testing timeline
methods
- Adds travis config
- Breaks up webpack configs into web and test configuration files that
inherit from a base config
