+++ .travis.yml
@@ -1,0 +1,19 @@
+---
+language: node_js
+node_js:
+  - "7"
+
+sudo: false
+
+cache:
+  directories:
+    - $HOME/.npm
+
+before_install:
+  - npm config set spin false
+
+install:
+  - npm install
+
+script:
+  - npm test
