+++ package.json
@@ -4,10 +4,11 @@
   "description": "Microsite to display events compiled by Lawyers for Good Government Volunteers.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "NODE_ENV=test webpack --hide-modules",
+    "test:watch": "NODE_ENV=test webpack --watch --hide-modules",
     "start": "webpack-dev-server --progress --colors --watch",
     "prod": "node server.js",
-    "build": "NODE_ENV=production webpack",
+    "build": "NODE_ENV=production webpack --display-error-details",
     "create": "node create-event.js",
     "heroku-postbuild": "npm run build"
   },
@@ -24,19 +25,24 @@
     "babel-preset-es2015-native-modules": "^6.9.4",
     "babel-preset-stage-2": "^6.22.0",
     "babili-webpack-plugin": "0.0.11",
+    "chai": "^3.5.0",
+    "chai-json-schema": "^1.4.0",
     "css-loader": "^0.26.2",
     "dir-loader": "^0.3.0",
     "file-loader": "^0.10.1",
     "html-loader": "^0.4.5",
     "html-webpack-plugin": "^2.28.0",
     "mkdirp": "^0.5.1",
+    "mocha": "^3.2.0",
     "node-sass": "^4.5.0",
     "raw-loader": "^0.5.1",
     "sass-loader": "^6.0.2",
     "script-loader": "^0.7.0",
     "style-loader": "^0.13.2",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.2.1"
+    "webpack": "^2.4.1",
+    "webpack-dev-server": "^2.2.1",
+    "webpack-node-externals": "^1.5.4",
+    "webpack-shell-plugin": "^0.5.0"
   },
   "dependencies": {
     "TimelineJS3": "github:L4GG/TimelineJS3",
