+++ src/test.js
@@ -1,0 +1,4 @@
+const context = require.context('../tests', true, /\.js$/)
+context.keys().forEach(context)
+
+module.exports = context
